@@ -1,57 +1,57 @@
 {
     "name": "sepes-web",
-    "version": "0.1.6",
+    "version": "0.1.16",
     "private": true,
     "dependencies": {
-        "@azure/storage-blob": "^12.1.2",
-        "@equinor/eds-core-react": "^0.8.3",
-        "@fortawesome/fontawesome-free": "^5.13.0",
-        "@fortawesome/fontawesome-svg-core": "^1.2.28",
-        "@fortawesome/free-solid-svg-icons": "^5.13.0",
-        "@fortawesome/react-fontawesome": "^0.1.9",
-        "@material-ui/core": "^4.10.0",
-        "@microsoft/applicationinsights-web": "^2.5.6",
+        "@azure/storage-blob": "^12.4.1",
+        "@equinor/eds-core-react": "^0.8.5",
+        "@fortawesome/fontawesome-free": "^5.15.2",
+        "@fortawesome/fontawesome-svg-core": "^1.2.34",
+        "@fortawesome/free-solid-svg-icons": "^5.15.2",
+        "@fortawesome/react-fontawesome": "^0.1.14",
+        "@material-ui/core": "^4.11.3",
+        "@microsoft/applicationinsights-web": "^2.5.11",
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^9.5.0",
         "@testing-library/user-event": "^7.2.1",
         "@types/jest": "^24.9.1",
-        "@types/node": "^12.12.37",
-        "@types/react": "^16.9.34",
-        "@types/react-dom": "^16.9.7",
-        "@types/react-router": "^5.1.7",
-        "@types/react-router-dom": "^5.1.5",
-        "@types/reactstrap": "^8.4.2",
-        "@types/styled-components": "^5.1.0",
-        "@typescript-eslint/eslint-plugin": "^4.8.2",
-        "@typescript-eslint/parser": "^4.8.2",
-        "animate.css": "^4.1.0",
+        "@types/node": "^12.19.15",
+        "@types/react": "^17.0.3",
+        "@types/react-dom": "^16.9.10",
+        "@types/react-router": "^5.1.11",
+        "@types/react-router-dom": "^5.1.7",
+        "@types/reactstrap": "^8.7.1",
+        "@types/styled-components": "^5.1.7",
+        "@typescript-eslint/eslint-plugin": "^4.15.0",
+        "@typescript-eslint/parser": "^4.15.0",
+        "animate.css": "^4.1.1",
         "axios": "^0.21.1",
-        "bootstrap": "^4.4.1",
-        "browser-image-compression": "^1.0.12",
-        "commitizen": "^4.1.2",
+        "bootstrap": "^4.6.0",
+        "browser-image-compression": "^1.0.14",
+        "commitizen": "^4.2.3",
         "copy-to-clipboard": "^3.3.1",
         "enzyme": "^3.11.0",
-        "enzyme-adapter-react-16": "^1.15.4",
+        "enzyme-adapter-react-16": "^1.15.6",
         "js-cookie": "^2.2.1",
         "lodash": ">=4.17.19",
         "lodash.debounce": "^4.0.8",
-        "msal": "^1.3.0",
+        "msal": "1.3.0",
         "node-forge": ">=0.10.0",
         "node-sass": "^4.14.1",
         "object-path": ">=0.11.5",
         "react": "^16.13.1",
-        "react-copy-to-clipboard": "^5.0.2",
+        "react-copy-to-clipboard": "^5.0.3",
         "react-dom": "^16.13.1",
-        "react-dropzone": "^11.0.1",
-        "react-notifications-component": "^2.4.0",
-        "react-router": "^5.1.2",
-        "react-router-dom": "^5.1.2",
-        "react-scripts": "3.4.4",
-        "react-select": "^3.1.0",
-        "reactstrap": "^8.4.1",
+        "react-dropzone": "^11.3.1",
+        "react-notifications-component": "^2.4.1",
+        "react-router": "^5.2.0",
+        "react-router-dom": "^5.2.0",
+        "react-scripts": "^4.0.3",
+        "react-select": "^3.2.0",
+        "reactstrap": "^8.9.0",
         "serialize-javascript": ">=3.1.0",
-        "styled-components": "^5.1.0",
-        "typescript": "^4.1.2"
+        "styled-components": "^5.2.1",
+        "typescript": "^4.2.3"
     },
     "scripts": {
         "start": "react-scripts -r @cypress/instrument-cra start",
