@@ -1,5 +1,6 @@
+/*eslint-disable consistent-return, no-shadow, react/no-array-index-key */
 import React, { useEffect, useState } from 'react';
-import { Button, Typography, Menu, DotProgress, Tooltip } from '@equinor/eds-core-react';
+import { Button, Typography, Menu, DotProgress, Tooltip, Icon } from '@equinor/eds-core-react';
 import { Link, useHistory } from 'react-router-dom';
 import styled from 'styled-components';
 import Stepper from '@material-ui/core/Stepper';
@@ -7,13 +8,15 @@ import Step from '@material-ui/core/Step';
 import StepLabel from '@material-ui/core/StepLabel';
 import DeleteResourceComponent from '../common/customComponents/DeleteResourceComponent';
 import { EquinorIcon } from '../common/StyledComponents';
-import { deleteSandbox, getResourceStatus, makeAvailable } from '../../services/Api';
+import { deleteSandbox, getResourceStatus, makeAvailable, getSandboxCostAnalysis } from '../../services/Api';
 import * as notify from '../common/notify';
 import { SandboxObj } from '../common/interfaces';
 import { getSandboxByIdUrl, getStudyByIdUrl } from '../../services/ApiCallStrings';
 import SureToProceed from '../common/customComponents/SureToProceed';
-import { resourceStatus, resourceType } from '../common/types';
-let set = require('lodash/set');
+import { resourceStatus, resourceType } from '../common/staticValues/types';
+import LoadingFull from '../common/LoadingComponentFullscreen';
+
+const set = require('lodash/set');
 
 const { MenuItem } = Menu;
 
@@ -31,12 +34,12 @@ const BtnTwoWrapper = styled.div`
     grid-template-columns: 160px 160px;
     grid-gap: 8px;
 `;
-
+/*
 const BtnThreeWrapper = styled.div`
     display: grid;
     grid-template-columns: 160px 160px 160px;
     grid-gap: 8px;
-`;
+`;*/
 
 type StepBarProps = {
     setStep: (value: any) => void;
@@ -53,6 +56,9 @@ type StepBarProps = {
     setLoading: any;
     setNewPhase: any;
     setDeleteSandboxInProgress: any;
+    setNewCostanalysisLink: any;
+    controller: AbortController;
+    vmsWithOpenInternet: any;
 };
 
 const getSteps = () => {
@@ -66,7 +72,11 @@ const getSteps = () => {
             label: 'Data sets available',
             description:
                 'Data sets become available in the sandbox, and you can now work your magic. Data set restrictions are applied.'
-        },
+        }
+    ];
+};
+/*
+,
         {
             label: 'Data retention',
             description:
@@ -76,8 +86,7 @@ const getSteps = () => {
             label: 'Decommission sandbox',
             description: 'Resources shuts down. Data will be removed according to your data retention  choices.'
         }
-    ];
-};
+*/
 
 let resourcesFailed = false;
 const interval = 20000; //20 seconds
@@ -96,13 +105,18 @@ const StepBar: React.FC<StepBarProps> = ({
     setUserClickedDelete,
     setLoading,
     setNewPhase,
-    setDeleteSandboxInProgress
+    setDeleteSandboxInProgress,
+    setNewCostanalysisLink,
+    controller,
+    vmsWithOpenInternet
 }) => {
     const history = useHistory();
     const steps = getSteps();
     const [userClickedMakeAvailable, setUserClickedMakeAvailable] = useState<boolean>(false);
     const [makeAvailableInProgress, setMakeAvailableInProgress] = useState<boolean>(false);
     const [allResourcesOk, setAllResourcesOk] = useState<boolean>(false);
+    const [sandboxHasVm, setSandboxHasVm] = useState<boolean>(false);
+    const [anyVmWithOpenInternet, setAnyVmWithOpenInternet] = useState<boolean>(false);
 
     useEffect(() => {
         getResources();
@@ -123,10 +137,10 @@ const StepBar: React.FC<StepBarProps> = ({
     }, [userClickedDelete]);
 
     const getResources = () => {
-        getResourceStatus(sandboxId).then((result: any) => {
+        getResourceStatus(sandboxId, controller.signal).then((result: any) => {
             if (result && (result.errors || result.Message)) {
                 resourcesFailed = true;
-                notify.show('danger', '500', result.Message, result.RequestId);
+                notify.show('danger', '500', result);
                 console.log('Err');
             } else {
                 setResources(result);
@@ -141,15 +155,35 @@ const StepBar: React.FC<StepBarProps> = ({
             return res;
         }
         let hasVm = false;
+        let noOpenInternet = true;
         resourcesIn.map((resource: any, i: number) => {
             if (resource.status !== resourceStatus.ok) {
                 res = false;
             }
             if (resource.type === resourceType.virtualMachine) {
                 hasVm = true;
+                if (resource.additionalProperties && resource.additionalProperties.InternetIsOpen) {
+                    noOpenInternet = false;
+                }
+            }
+            if (resource.type === resourceType.resourceGroup && sandbox.linkToCostAnalysis === null) {
+                getCostAnalysisLinkToSandbox();
             }
         });
-        setAllResourcesOk(res && hasVm);
+        setAnyVmWithOpenInternet(!noOpenInternet);
+        setSandboxHasVm(hasVm);
+        setAllResourcesOk(res && hasVm && noOpenInternet);
+    };
+
+    const getCostAnalysisLinkToSandbox = () => {
+        getSandboxCostAnalysis(sandboxId).then((result: any) => {
+            if (result && result.Message) {
+                notify.show('danger', '500', result);
+            } else {
+                setNewCostanalysisLink(result);
+                setSandbox({ ...sandbox, linkToCostAnalysis: result });
+            }
+        });
     };
 
     const [state, setState] = React.useState<{
@@ -181,9 +215,9 @@ const StepBar: React.FC<StepBarProps> = ({
         setUpdateCache({ ...updateCache, [getStudyByIdUrl(studyId)]: true });
         deleteSandbox(sandboxId).then((result: any) => {
             setLoading(false);
-            if (result.Message) {
+            if (result && result.Message) {
                 setDeleteSandboxInProgress(false);
-                notify.show('danger', '500', result.Message, result.RequestId);
+                notify.show('danger', '500', result);
             }
             history.push('/studies/' + studyId);
         });
@@ -197,9 +231,10 @@ const StepBar: React.FC<StepBarProps> = ({
             setMakeAvailableInProgress(false);
             if (result.Message || result.errors) {
                 setNewPhase(0);
-                notify.show('danger', '500', result.Message, result.RequestId);
+                notify.show('danger', '500', result);
             } else {
                 setSandbox(set({ ...sandbox }, 'permissions.openInternet', result.permissions.openInternet));
+                setSandbox(set({ ...sandbox }, 'datasets', result.datasets));
                 setNewPhase(1);
             }
         });
@@ -209,22 +244,41 @@ const StepBar: React.FC<StepBarProps> = ({
         if (sandbox.permissions && !sandbox.permissions.increasePhase) {
             return 'You do not have permission to make this sandbox Available';
         }
-        if (!allResourcesOk || sandbox.datasets.length === 0) {
-            return 'All resources must have status OK and atleast one VM and Data set must be in the sandbox';
+        if (!sandboxHasVm) {
+            return 'You need atleast one VM in the sandbox';
+        }
+        if (anyVmWithOpenInternet) {
+            return 'One or more vms have open internet. Close before making sandbox available';
+        }
+        if (sandbox.datasets.length === 0) {
+            return 'No datasets in the sandbox';
+        }
+        if (!allResourcesOk) {
+            return 'All resources must have status OK';
         }
         return '';
     };
 
     const optionsTemplate = (
         <>
-            <MenuItem
-                onClick={() => setUserClickedDelete(true)}
-                data-cy="sandbox_delete"
-                disabled={sandbox.permissions && !sandbox.permissions.delete}
+            <Tooltip
+                title={
+                    sandbox.permissions && !sandbox.permissions.delete
+                        ? 'You do not have access to delete this sandbox'
+                        : ''
+                }
+                placement="left"
+                open={sandbox.permissions && !sandbox.permissions.delete}
             >
-                {EquinorIcon('delete_forever', 'red', 24)}
-                <span style={{ color: 'red' }}>Delete sandbox</span>
-            </MenuItem>
+                <MenuItem
+                    onClick={() => setUserClickedDelete(true)}
+                    data-cy="sandbox_delete"
+                    disabled={sandbox.permissions && !sandbox.permissions.delete}
+                >
+                    {EquinorIcon('delete_forever', 'red', 24)}
+                    <span style={{ color: 'red' }}>Delete sandbox</span>
+                </MenuItem>
+            </Tooltip>
         </>
     );
 
@@ -240,7 +294,7 @@ const StepBar: React.FC<StepBarProps> = ({
                     onClick={(e) => (isOpen ? closeMenu() : openMenu(e, 'first'))}
                     data-cy="sandbox_more_options"
                 >
-                    More options
+                    <span style={{ marginLeft: '0' }}>More options</span>
                     {EquinorIcon('more_vertical', '#007079', 24)}
                 </Button>
                 <Menu
@@ -277,20 +331,13 @@ const StepBar: React.FC<StepBarProps> = ({
                                             sandbox.permissions.increasePhase &&
                                             sandbox.datasets.length > 0 &&
                                             !makeAvailableInProgress &&
-                                            allResourcesOk
+                                            allResourcesOk &&
+                                            !vmsWithOpenInternet
                                         )
                                     }
                                 >
-                                    {makeAvailableInProgress ? (
-                                        <div>
-                                            <DotProgress variant="green" />
-                                        </div>
-                                    ) : (
-                                        <>
-                                            <span>Make available</span>
-                                            {EquinorIcon('arrow_forward', '#FFFFFF', 16, () => {}, true)}
-                                        </>
-                                    )}
+                                    <span style={{ marginLeft: '0' }}>Make available</span>
+                                    {EquinorIcon('chevron_right', '#FFFFFF', 24, () => {}, true)}
                                 </Button>
                             </Tooltip>
                         </div>
@@ -311,7 +358,7 @@ const StepBar: React.FC<StepBarProps> = ({
                         </Button>
                         */}
                         {returnOptionsButton()}
-                        {/* 
+                        {/*
                         <Button
                             onClick={() => {
                                 setStep(2);
@@ -332,7 +379,7 @@ const StepBar: React.FC<StepBarProps> = ({
                                 setStep(0);
                             }}
                         >
-                            {EquinorIcon('arrow_back', '#007079', 16, () => {}, true)}Make Available
+                            {EquinorIcon('chevron_left', '#007079', 16, () => {}, true)}Make Available
                         </Button>
                         {returnOptionsButton()}
                     </BtnTwoWrapper>
@@ -341,64 +388,89 @@ const StepBar: React.FC<StepBarProps> = ({
         }
     };
     return (
-        <Wrapper>
-            <div>
-                <Link to={'/studies/' + studyId} style={{ color: '#007079', fontSize: '22px', margin: '0 0 0 16px' }}>
-                    {EquinorIcon('arrow_back', '#007079', 24, () => {}, true)}
-                </Link>
-                <Typography style={{ display: 'inline-block', marginLeft: '16px' }} variant="h2">
-                    {sandbox && sandbox.name}
-                </Typography>
-                <div style={{ float: 'right' }}>{returnControlButtons()}</div>
-            </div>
-            <Stepper activeStep={step} alternativeLabel nonLinear>
-                {steps.map((stepL: any, index) => {
-                    const stepProps = {};
-                    const labelProps: any = {};
-                    labelProps.optional = <Typography variant="caption">{stepL.description}</Typography>;
-
-                    return (
-                        <Step key={index}>
-                            <StepLabel {...labelProps}>{stepL.label}</StepLabel>
-                        </Step>
-                    );
-                })}
-            </Stepper>
-            <div style={{ float: 'right' }}>
-                <a
-                    href={sandbox.linkToCostAnalysis}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    style={{
-                        color: '#007079',
-                        fontSize: '22px',
-                        margin: '0 0 0 16px',
-                        display: 'inline-block',
-                        float: 'right'
-                    }}
-                >
-                    <Typography style={{ display: 'inline-block', marginRight: '8px', fontSize: '16px' }} variant="h2">
-                        Cost analysis
+        <>
+            {makeAvailableInProgress && <LoadingFull noTimeout />}
+            <Wrapper>
+                <div style={{ display: 'flex' }}>
+                    <Link to={'/studies/' + studyId} style={{ color: '#007079', fontSize: '22px' }}>
+                        <Button variant="ghost_icon">
+                            <Icon
+                                style={{ marginBottom: '' }}
+                                color="#007079"
+                                name="chevron_left"
+                                size={24}
+                                title="chevron_left"
+                            />
+                        </Button>
+                    </Link>
+                    <Typography style={{ display: 'inline-block', marginLeft: '8px' }} variant="h2">
+                        {sandbox && sandbox.name}
                     </Typography>
-                    {EquinorIcon('external_link', '#007079', 24, () => {}, true)}
-                </a>
-            </div>
-            {userClickedDelete && (
-                <DeleteResourceComponent
-                    ResourceName={sandbox.name}
-                    setUserClickedDelete={setUserClickedDelete}
-                    onClick={deleteThisSandbox}
-                    type="sandbox"
-                />
-            )}
-            {userClickedMakeAvailable && (
-                <SureToProceed
-                    setUserClickedButton={setUserClickedMakeAvailable}
-                    onClick={makeThisSandboxAvailable}
-                    type="making this sandbox available"
-                />
-            )}
-        </Wrapper>
+                    <div style={{ marginLeft: 'auto' }}>{returnControlButtons()}</div>
+                </div>
+                <Stepper activeStep={step} alternativeLabel nonLinear color="red">
+                    {steps.map((stepL: any, index) => {
+                        const labelProps: any = {};
+                        labelProps.optional = <Typography variant="caption">{stepL.description}</Typography>;
+
+                        return (
+                            <Step key={index} style={{ padding: '0 96px 0 96px' }}>
+                                <StepLabel {...labelProps}>
+                                    <span style={{ textAlign: 'center' }}>{stepL.label}</span>
+                                </StepLabel>
+                            </Step>
+                        );
+                    })}
+                </Stepper>
+
+                <div style={{ marginLeft: 'auto' }}>
+                    <Tooltip
+                        title={sandbox.linkToCostAnalysis ? '' : 'Link will be available when resource group is ready'}
+                        placement="left"
+                    >
+                        <a
+                            href={sandbox.linkToCostAnalysis}
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            style={{
+                                color: '#007079',
+                                fontSize: '22px',
+                                margin: '0 0 0 16px',
+                                display: 'inline-block',
+                                float: 'right'
+                            }}
+                        >
+                            <Typography
+                                style={{ display: 'inline-block', marginRight: '8px', fontSize: '16px' }}
+                                variant="h2"
+                            >
+                                Cost analysis
+                            </Typography>
+                            {sandbox.linkToCostAnalysis ? (
+                                EquinorIcon('external_link', '#007079', 24, () => {}, true)
+                            ) : (
+                                <DotProgress color="primary" />
+                            )}
+                        </a>
+                    </Tooltip>
+                </div>
+                {userClickedDelete && (
+                    <DeleteResourceComponent
+                        ResourceName={sandbox.name}
+                        setUserClickedDelete={setUserClickedDelete}
+                        onClick={deleteThisSandbox}
+                        type="sandbox"
+                    />
+                )}
+                {userClickedMakeAvailable && (
+                    <SureToProceed
+                        setUserClickedButton={setUserClickedMakeAvailable}
+                        onClick={makeThisSandboxAvailable}
+                        type="making the selected data sets available to this sandbox"
+                    />
+                )}
+            </Wrapper>
+        </>
     );
 };
 
