@@ -1,6 +1,6 @@
 import React from 'react';
 import Bar from './topbar/TopBar';
-import ErrorBoundary from '../components/common/ErrorBoundary';
+import ErrorBoundary from './common/ErrorBoundary';
 import ReactNotification from 'react-notifications-component';
 import 'react-notifications-component/dist/theme.css';
 
@@ -14,6 +14,6 @@ const Layout = (props: any) => {
             </ErrorBoundary>
         </>
     );
-}
+};
 
 export default Layout;
