@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { useHistory } from 'react-router-dom';
-import { Prompt as ReactPrompt } from 'react-router';
+import { useHistory, Prompt as ReactPrompt } from 'react-router-dom';
 import { Dialog, Button, Scrim } from '@equinor/eds-core-react';
 import styled from 'styled-components';
 
@@ -17,9 +16,10 @@ const TempButtonWrapper = styled.div`
 type PromptProps = {
     hasChanged: boolean;
     fallBackAddress?: string;
+    customText?: string;
 };
 
-const Prompt: React.FC<PromptProps> = ({ hasChanged, fallBackAddress }) => {
+const Prompt: React.FC<PromptProps> = ({ hasChanged, fallBackAddress, customText }) => {
     const history = useHistory();
     const [visibleScrim, setVisibleScrim] = useState<boolean>(false);
     const [confirmedNavigation, setConfirmedNavigation] = useState<boolean>(false);
@@ -48,7 +48,7 @@ const Prompt: React.FC<PromptProps> = ({ hasChanged, fallBackAddress }) => {
                         <Title>Unsaved changes</Title>
                         <CustomContent scrollable={false}>
                             Are you sure you want to leave this page? <br />
-                            All unsaved changes will be lost.
+                            {customText || 'All unsaved changes will be lost.'}
                         </CustomContent>
                         <span style={{ marginLeft: 'auto' }}>
                             <Actions>
