@@ -1,8 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { Divider } from '@equinor/eds-core-react';
-import { Icon } from '@equinor/eds-core-react';
+import { Icon, Divider } from '@equinor/eds-core-react';
 import {
     cloud_upload,
     delete_forever,
@@ -15,7 +14,8 @@ import {
     key,
     check,
     error_outlined,
-    more_vertical
+    more_vertical,
+    refresh
 } from '@equinor/eds-icons';
 
 const icons = {
@@ -30,7 +30,8 @@ const icons = {
     key,
     check,
     error_outlined,
-    more_vertical
+    more_vertical,
+    refresh
 };
 Icon.add(icons);
 
@@ -116,7 +117,7 @@ const StyledRequiredLabel = styled.span`
 `;
 
 export const Title = (props: any) => {
-    let { title } = props;
+    const { title } = props;
 
     return (
         <div>
@@ -127,7 +128,7 @@ export const Title = (props: any) => {
 };
 
 export const TextArea = (props: any) => {
-    let { label, text, textStyle } = props;
+    const { label, text, textStyle } = props;
 
     return (
         <StyledTextArea {...props}>
@@ -138,7 +139,7 @@ export const TextArea = (props: any) => {
 };
 
 export const TextAreaDescription = (props: any) => {
-    let { label, text, textStyle } = props;
+    const { label, text, textStyle } = props;
 
     return (
         <StyledTextArea {...props}>
@@ -149,7 +150,7 @@ export const TextAreaDescription = (props: any) => {
 };
 
 export const RequiredLabel = (props: any) => {
-    let { label } = props;
+    const { label } = props;
 
     return (
         <StyledRequiredLabel {...props}>
