@@ -9,17 +9,16 @@ import {
     createStandardDataset,
     updateStandardDataset
 } from '../../services/Api';
-import { checkIfRequiredFieldsAreNull } from '../common/helpers';
-import { useHistory } from 'react-router-dom';
+import { checkIfInputIsNumberWihoutCharacters, checkIfRequiredFieldsAreNull } from '../common/helpers';
 import * as notify from '../common/notify';
 import Promt from '../common/Promt';
 import { UpdateCache } from '../../App';
 import { EquinorIcon } from '../common/StyledComponents';
 import { Permissions } from '../../index';
-import NoAccess from '../common/NoAccess';
-import { useLocation } from 'react-router-dom';
+import NoAccess from '../common/informationalComponents/NoAccess';
+import { useLocation, useHistory } from 'react-router-dom';
 import useFetchUrl from '../common/hooks/useFetchUrl';
-import { dataInventoryLink, ClassificationGuidlinesLink } from '../common/commonLinks';
+import { dataInventoryLink, ClassificationGuidlinesLink } from '../common/staticValues/commonLinks';
 import {
     getDatasetsInStudyUrl,
     getDatasetsUrl,
@@ -86,7 +85,6 @@ type CreateEditDatasetProps = {
     setDatasetFromDetails: (value: any) => void;
     setShowEditDataset: (value: any) => void;
     editingDataset: boolean;
-    cache: any;
     permissions: DatasetPermissionObj;
 };
 
@@ -95,7 +93,6 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
     setDatasetFromDetails,
     setShowEditDataset,
     editingDataset,
-    cache,
     permissions
 }) => {
     const studyId = window.location.pathname.split('/')[2];
@@ -169,8 +166,9 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
                     });
                     history.push('/studies/' + studyId + '/datasets/' + result.id);
                 } else {
+                    setUserPressedCreate(false);
                     console.log('Err');
-                    notify.show('danger', '500', result.Message, result.RequestId);
+                    notify.show('danger', '500', result);
                 }
             });
         } else if (isDatasetspecificDataset) {
@@ -181,12 +179,13 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
                     setUpdateCache({
                         ...updateCache,
                         [getStudyByIdUrl(studyId)]: true,
-                        [getStudySpecificDatasetUrl(result.Id, studyId)]: true
+                        [getStudySpecificDatasetUrl(result.id, result.studyId)]: true
                     });
                     setDatasetFromDetails(result);
                     setShowEditDataset(false);
                 } else {
-                    notify.show('danger', '500', result.Message, result.RequestId);
+                    setUserPressedCreate(false);
+                    notify.show('danger', '500', result);
                     console.log('Err');
                 }
             });
@@ -202,7 +201,8 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
                     });
                     history.push('/datasets/' + result.id);
                 } else {
-                    notify.show('danger', '500', result.Message, result.RequestId);
+                    setUserPressedCreate(false);
+                    notify.show('danger', '500', result);
                     console.log('Err');
                 }
             });
@@ -211,19 +211,30 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
                 setLoading(false);
                 if (result && !result.Message) {
                     setHasChanged(false);
-                    setUpdateCache({ ...updateCache, 'datasets/': true });
-                    cache[getStandardDatasetUrl(studyId)] = result;
+                    setUpdateCache({ ...updateCache, 'datasets/': true, [getStandardDatasetUrl(studyId)]: true });
                     setDatasetFromDetails(result);
                     setShowEditDataset(false);
                 } else {
-                    notify.show('danger', '500', result.Message, result.RequestId);
+                    setUserPressedCreate(false);
+                    notify.show('danger', '500', result);
                     console.log('Err');
                 }
             });
         }
     };
 
-    const handleChange = (columName: string, value: string) => {
+    const handleChange = (columName: string, value: any) => {
+        if (columName === 'dataId') {
+            if (value < 0 || value === '') {
+                setDataset({ ...dataset, dataId: undefined });
+            } else {
+                setDataset({
+                    ...dataset,
+                    dataId: parseInt(value)
+                });
+            }
+            return;
+        }
         setHasChanged(true);
         setDataset({
             ...dataset,
@@ -259,6 +270,9 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
         if (!dataset?.name?.length || !dataset?.classification?.length || !dataset?.location?.length) {
             return true;
         }
+        if (dataset?.dataId && !checkIfInputIsNumberWihoutCharacters(dataset?.dataId.toString())) {
+            return true;
+        }
         return false;
     };
 
@@ -271,7 +285,7 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
         );
     };
 
-    return (checkUrlIfGeneralDataset && generalDatasetpermissions.canEdit_PreApproved_Datasets) ||
+    return (checkUrlIfGeneralDataset() && generalDatasetpermissions.canEdit_PreApproved_Datasets) ||
         (permissions && permissions.editDataset) ||
         (location && location.state.canCreateStudySpecificDataset) ? (
         <>
@@ -312,7 +326,7 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
                             data-cy="dataset_storage_name"
                             inputIcon={
                                 <div style={{ position: 'relative', right: '4px', bottom: '4px' }}>
-                                    <Tooltip title="This cannot be changed later" placement={'right'}>
+                                    <Tooltip title="This cannot be changed later" placement="right">
                                         {EquinorIcon('error_outlined', '#6F6F6F', 24)}
                                     </Tooltip>
                                 </div>
@@ -378,7 +392,7 @@ const CreateEditDataset: React.FC<CreateEditDatasetProps> = ({
                     </StyledLink>
                     <SaveCancelWrapper>
                         <Button disabled={checkForInputErrors() || loading} onClick={addDataset} data-cy="dataset_save">
-                            {loading ? <DotProgress variant="green" /> : 'Save'}
+                            {loading ? <DotProgress color="primary" /> : 'Save'}
                         </Button>
                         <Button disabled={userPressedCreate || loading} onClick={handleCancel} variant="outlined">
                             Cancel
