@@ -1,33 +1,38 @@
-import React, { useState, useContext } from 'react';
+/*eslint-disable consistent-return, no-unneeded-ternary */
+import React, { useState, useContext, useEffect } from 'react';
 import styled from 'styled-components';
-import { Typography, Icon, Button, Tooltip, LinearProgress } from '@equinor/eds-core-react';
-import { DatasetObj } from '../common/interfaces';
-import { deleteFileInDataset, removeStudyDataset } from '../../services/Api';
-import { Link } from 'react-router-dom';
+import { Typography, Icon, Button, Tooltip, LinearProgress, DotProgress, Chip } from '@equinor/eds-core-react';
+import { DatasetObj, DatasetResourcesObj } from '../common/interfaces';
+import {
+    getDatasetSasToken,
+    getStudySpecificDatasetFiles,
+    getStudySpecificDatasetResources,
+    removeStudyDataset,
+    getDatasetSasTokenDelete
+} from '../../services/Api';
 import { arrow_back, delete_forever } from '@equinor/eds-icons';
-import { Label } from '../common/StyledComponents';
-import { bytesToMB } from '../common/helpers';
+import { bytesToSize } from '../common/helpers';
 import LoadingFull from '../common/LoadingComponentFullscreen';
 import CreateEditDataset from './CreateEditDataset';
 import Dropzone from '../common/upload/DropzoneFile';
-import { loginRequest, makeFileBlobFromUrl } from '../../auth/AuthFunctions';
+import { makeFileBlobFromUrl } from '../../auth/AuthFunctions';
 import { Permissions } from '../../index';
-import { useLocation } from 'react-router-dom';
 import useFetchUrl from '../common/hooks/useFetchUrl';
 import * as notify from '../common/notify';
-import { EquinorIcon } from '../common/StyledComponents';
-import { useHistory } from 'react-router-dom';
+import { EquinorIcon, Label } from '../common/StyledComponents';
+import { useHistory, Link } from 'react-router-dom';
 import DeleteResourceComponent from '../common/customComponents/DeleteResourceComponent';
 import { UpdateCache } from '../../App';
-import myMSALObj from '../../auth/AuthConfig';
 import {
     getDatasetsFilesUrl,
     getStandardDatasetUrl,
     getStudySpecificDatasetUrl,
     getStudyByIdUrl
 } from '../../services/ApiCallStrings';
-import NotFound from '../common/NotFound';
-import axios from 'axios';
+import NotFound from '../common/informationalComponents/NotFound';
+import { resourceStatus, resourceType } from '../common/staticValues/types';
+import { uploadFile, deleteFile } from '../../services/BlobStorage';
+import Prompt from '../common/Promt';
 
 const icons = {
     arrow_back,
@@ -72,98 +77,233 @@ const checkUrlIfGeneralDataset = () => {
     }
     return false;
 };
+let controller = new AbortController();
+let controllerFiles = new AbortController();
+let controllerSas = new AbortController();
+const interval = 7000;
+const intervalUpdateSas = 1740000;
+const intervalUpdateSasDelete = 285000;
 
-let cancelToken = axios.CancelToken;
-let source = cancelToken.source();
+let abortArray: any = [];
+let progressArray: any = [];
 
 const DatasetDetails = (props: any) => {
-    let datasetId = window.location.pathname.split('/')[4];
+    const datasetId = window.location.pathname.split('/')[4];
     const studyId = window.location.pathname.split('/')[2];
     const isStandard = checkUrlIfGeneralDataset();
     const [userClickedDelete, setUserClickedDelete] = useState<boolean>(false);
     const [datasetDeleteInProgress, setDatasetDeleteInProgress] = useState<boolean>(false);
     const [loading, setLoading] = useState<boolean>(false);
+    const [loadingFiles, setLoadingFiles] = useState<boolean>(false);
+    const [isSubscribed, setIsSubscribed] = useState<boolean>(true);
     const [dataset, setDataset] = useState<DatasetObj>({
         name: '',
+        storageAccountLink: undefined,
         permissions: {
             deleteDataset: false,
             editDataset: false
         }
     });
-    useFetchUrl(
+
+    const datasetResponse = useFetchUrl(
         isStandard ? getStandardDatasetUrl(studyId) : getStudySpecificDatasetUrl(datasetId, studyId),
-        setDataset
+        setDataset,
+        undefined,
+        controller
     );
     const [showEditDataset, setShowEditDataset] = useState<boolean>(false);
+    const [duplicateFiles, setDuplicateFiles] = useState<boolean>(false);
+    const [hasChanged, setHasChanged] = useState<boolean>(false);
     const [files, setFiles] = useState<any>([]);
-    const datasetResponse = useFetchUrl(
-        isStandard ? getDatasetsFilesUrl(studyId) : getDatasetsFilesUrl(datasetId),
-        setFiles,
-        !isStandard
+    const [datasetStorageAccountIsReady, setDatasetStorageAccountIsReady] = useState<Boolean>(
+        dataset.storageAccountLink !== '' || false
     );
     const permissions = useContext(Permissions);
     const { updateCache, setUpdateCache } = useContext(UpdateCache);
     const history = useHistory();
-    const [percentComplete, setPercentComplete] = useState<any>(0);
-    let keyCount: number = 0;
+    const [storageAccountStatus, setStorageAccountStatus] = useState<string>('');
+    const [sasKey, setSasKey] = useState<string>('');
+    const [sasKeyExpired, setSasKeyExpired] = useState<boolean>(true);
+    const [sasKeyDelete, setSasKeyDelete] = useState<string>('');
+    const [sasKeyDeleteExpired, setSasKeyDeleteExpired] = useState<boolean>(true);
 
-    const getKey = () => {
-        return keyCount++;
+    useEffect(() => {
+        let timer: any;
+        try {
+            timer = setInterval(async () => {
+                if (dataset.storageAccountLink === '' || dataset.storageAccountLink === null) {
+                    getDatasetResources();
+                }
+            }, interval);
+        } catch (e) {
+            console.log(e);
+        }
+
+        return () => {
+            clearInterval(timer);
+            setIsSubscribed(false);
+        };
+    }, [dataset.storageAccountLink]);
+
+    useEffect(() => {
+        if (
+            dataset.storageAccountLink !== '' &&
+            dataset.storageAccountLink !== undefined &&
+            dataset.storageAccountLink !== null
+        ) {
+            setIsSubscribed(true);
+            getDatasetFiles();
+        }
+        return () => setIsSubscribed(false);
+    }, [datasetStorageAccountIsReady, dataset]);
+
+    useEffect(() => {
+        const filesInProgress = progressArray.filter((x) => x.percent && x.percent > 0 && x.percent < 100);
+
+        if (filesInProgress.length > 0) {
+            setHasChanged(true);
+        } else {
+            setHasChanged(false);
+        }
+    }, [files, progressArray]);
+
+    useEffect(() => {
+        return () => {
+            cancelGettingFilesCall();
+            cancelAllDownloads();
+            abortArray = [];
+        };
+    }, []);
+
+    useEffect(() => {
+        const timer = setInterval(async () => {
+            setSasKeyExpired(true);
+        }, intervalUpdateSas);
+
+        return () => clearInterval(timer);
+    }, []);
+
+    useEffect(() => {
+        const timer = setInterval(async () => {
+            setSasKeyDeleteExpired(true);
+        }, intervalUpdateSasDelete);
+
+        return () => clearInterval(timer);
+    }, []);
+
+    const getSasKey = (retries = 3, backoff = 300) => {
+        return new Promise((resolve) => {
+            if (!sasKeyExpired) {
+                return resolve(sasKey);
+            }
+            getDatasetSasToken(datasetId, controllerSas.signal)
+                .then((result: any) => {
+                    if (retries > 0 && result.Message) {
+                        setTimeout(() => {
+                            /* 2 */
+                            return getSasKey(retries - 1); /* 3 */
+                        }, backoff);
+                    }
+
+                    setSasKeyExpired(false);
+                    setSasKey(result);
+                    return resolve(result);
+                })
+                .catch((ex: any) => {
+                    console.log(ex);
+                    if (retries > 0) {
+                        setTimeout(() => {
+                            /* 2 */
+                            return getSasKey(retries - 1); /* 3 */
+                        }, backoff);
+                    }
+                });
+        });
     };
 
-    const uploadFiles = (formData: any, previousFiles: any) => {
-        setPercentComplete(0);
-        updateOnNextVisit();
+    const getSasKeyDelete = (retries = 3, backoff = 300) => {
+        return new Promise((resolve) => {
+            if (!sasKeyDeleteExpired) {
+                return resolve(sasKeyDelete);
+            }
+            getDatasetSasTokenDelete(datasetId, controllerSas.signal)
+                .then((result: any) => {
+                    if (retries > 0 && result.Message) {
+                        setTimeout(() => {
+                            /* 2 */
+                            return getSasKeyDelete(retries - 1); /* 3 */
+                        }, backoff);
+                    }
+                    setSasKeyDeleteExpired(false);
+                    setSasKeyDelete(result);
+                    return resolve(result);
+                })
+                .catch((ex: any) => {
+                    console.log(ex);
+                    if (retries > 0) {
+                        setTimeout(() => {
+                            /* 2 */
+                            return getSasKey(retries - 1); /* 3 */
+                        }, backoff);
+                    }
+                });
+        });
+    };
+
+    const cancelGettingFilesCall = (): void => {
+        controllerFiles.abort();
+        controllerFiles = new AbortController();
+    };
+
+    const getDatasetResources = () => {
         if (!checkUrlIfGeneralDataset()) {
-            postFile('api/datasets/' + datasetId + '/files', formData, previousFiles).then((result: any) => {
-                if (result.Message) {
-                    console.log('err', result);
-                }
-            });
-        } else {
-            datasetId = studyId;
-            postFile('api/datasets/' + datasetId + '/files', formData, previousFiles).then((result: any) => {
-                if (result.Message) {
-                    console.log('err', result);
+            getStudySpecificDatasetResources(datasetId, studyId).then((result: any) => {
+                if (result && (result.errors || result.Message)) {
+                    notify.show('danger', '500', result);
+                    console.log('Err');
+                } else {
+                    checkStatusOfStorageAccount(result);
                 }
             });
         }
     };
 
-    const postFile = async (url, files: any, previousFiles: any) => {
-        return new Promise(() => {
-            myMSALObj
-                .acquireTokenSilent(loginRequest)
-                .then((tokenResponse: any) => {
-                    if (tokenResponse.accessToken) {
-                        const bearer = `Bearer ${tokenResponse.accessToken}`;
-
-                        axios({
-                            headers: { Authorization: bearer },
-                            method: 'post',
-                            url: `${process.env.REACT_APP_SEPES_BASE_API_URL}${url}`,
-                            data: files,
-                            onUploadProgress: (p) => {
-                                const percentCalculated = Math.floor((p.loaded * 100) / p.total);
-                                setPercentComplete(percentCalculated);
-                            },
-                            cancelToken: source.token
-                        }).catch((thrown) => {
-                            if (axios.isCancel(thrown)) {
-                                setFiles(previousFiles);
-                            }
-                        });
-                    }
-
-                    // Callback code here
-                })
-                .catch((error: string) => {
-                    console.log(error);
-                });
-        });
+    const getDatasetFiles = () => {
+        if (!checkUrlIfGeneralDataset() && isSubscribed) {
+            setLoadingFiles(true);
+            getStudySpecificDatasetFiles(datasetId, controllerFiles.signal).then((result: any) => {
+                setLoadingFiles(false);
+                if (result && (result.errors || result.Message)) {
+                    notify.show('danger', '500', result);
+                    console.log('Err');
+                } else if (result && isSubscribed) {
+                    setFiles(result);
+                    progressArray = result;
+                }
+            });
+        }
     };
 
-    const handleEditMetdata = (evt) => {
+    const checkStatusOfStorageAccount = (resources: any) => {
+        let res = false;
+        if (!resources && !Array.isArray(resources)) {
+            return res;
+        }
+        resources.map((resource: DatasetResourcesObj) => {
+            setStorageAccountStatus(resource.status);
+            if (resource.status === resourceStatus.ok && resource.type === resourceType.storageAccount) {
+                res = true;
+                setDataset({ ...dataset, storageAccountLink: resource.linkToExternalSystem });
+                const dataCache = isStandard
+                    ? getStandardDatasetUrl(studyId)
+                    : getStudySpecificDatasetUrl(datasetId, studyId);
+                setUpdateCache({ ...updateCache, [dataCache]: true });
+            }
+        });
+        setDatasetStorageAccountIsReady(res);
+    };
+
+    const handleEditMetdata = () => {
         setShowEditDataset(true);
     };
 
@@ -173,6 +313,8 @@ const DatasetDetails = (props: any) => {
     };
 
     const deleteDataset = () => {
+        controllerFiles.abort();
+        controllerFiles = new AbortController();
         setLoading(true);
         setDatasetDeleteInProgress(true);
         setUserClickedDelete(false);
@@ -183,227 +325,353 @@ const DatasetDetails = (props: any) => {
                 history.push('/studies/' + studyId);
             } else {
                 console.log('Err');
-                notify.show('danger', '500', result.Message, result.RequestId);
+                notify.show('danger', '500', result);
             }
+        });
+    };
+
+    const setFilesProgressToOnePercent = (_files: any) => {
+        _files.forEach(async (file: any) => {
+            const filePercent = { blobName: file.name, percent: 1, controller: new AbortController() };
+            abortArray.push(filePercent);
         });
     };
 
     const handleFileDrop = async (_files: File[]): Promise<void> => {
+        setDuplicateFiles(false);
+        _files = checkIfFileAlreadyIsUploaded(_files);
+
+        if (_files.length === 0) {
+            return;
+        }
+
+        setHasChanged(true);
         const previousFiles = [...files];
         const tempFiles = [...files];
         tempFiles.push(..._files);
+
+        _files.forEach((_file: any) => {
+            _file.percent = 1;
+            progressArray.push(_file);
+        });
+
         setFiles(tempFiles);
-        let _formData = new FormData();
         if (_files.length) {
-            let filesHandledCount = 0;
-            await _files.forEach(async (file) => {
-                await makeFileBlobFromUrl(URL.createObjectURL(file), file.name)
-                    .then((blob) => {
-                        filesHandledCount++;
-                        _formData.append(`files`, blob);
-                    })
-                    .then(() => {
-                        if (filesHandledCount === _files.length) {
-                            uploadFiles(_formData, previousFiles);
-                        }
+            setFilesProgressToOnePercent(_files);
+            getSasKey().then((result: any) => {
+                if (result && !result.Message) {
+                    _files.forEach(async (file) => {
+                        await makeFileBlobFromUrl(URL.createObjectURL(file), file.name)
+                            .then((blob) => {
+                                try {
+                                    uploadFile(
+                                        result || sasKey,
+                                        file.name,
+                                        blob,
+                                        file.size,
+                                        abortArray,
+                                        setFiles,
+                                        progressArray
+                                    );
+                                } catch (ex) {
+                                    console.log(ex);
+                                    setFiles(previousFiles);
+                                }
+                            })
+                            .catch((ex) => {
+                                console.log(ex);
+                            });
                     });
+                } else {
+                    setFiles(previousFiles);
+                    console.log('Err');
+                    notify.show('danger', '500', result);
+                }
             });
         }
     };
 
+    const checkIfFileAlreadyIsUploaded = (_files) => {
+        const newArray: any = [];
+        _files.forEach((file: File) => {
+            const res = files
+                .map((e) => {
+                    return e.name;
+                })
+                .indexOf(file.name);
+            if (res === -1) newArray.push(file);
+        });
+
+        if (_files.length !== newArray.length) {
+            setDuplicateFiles(true);
+        }
+
+        return newArray;
+    };
+
     const removeFile = (i: number, file: any): void => {
-        setPercentComplete(0);
+        try {
+            controller.abort();
+            controller = new AbortController();
+        } catch (e) {
+            if (e.name === 'AbortError') {
+                // abort was called on our abortSignal
+                console.log('Operation was aborted by the user');
+            } else {
+                // some other error occurred 🤷‍♂️
+                console.log('Uploading file failed');
+            }
+        }
+
         updateOnNextVisit();
         const _files = [...files];
         _files.splice(i, 1);
         setFiles(_files);
-        deleteFileInDataset(datasetId, file.name).then((result: any) => {
-            if (result.Message) {
-                notify.show('danger', '500', result.Message, result.RequestId);
+        const index = abortArray.findIndex((x) => x.blobName === file.name);
+        const progIndex = progressArray.findIndex((x) => x.name === file.name);
+        if (progIndex !== -1) {
+            progressArray.splice(progIndex, 1);
+        }
+
+        if (index !== -1) {
+            const progressItem = abortArray[index];
+            if (progressItem && progressItem.percent === 1) {
+                try {
+                    controllerSas.abort();
+                    controllerSas = new AbortController();
+                    abortArray.splice(index, 1);
+                    return;
+                } catch (error) {
+                    console.log(error);
+                }
+            } else if (progressItem.percent < 100) {
+                try {
+                    progressItem.controller.abort();
+                } catch (error) {
+                    console.log(error);
+                }
+
+                abortArray.splice(index, 1);
+                return;
             }
-        });
+        }
+        getSasKeyDelete()
+            .then((result: any) => {
+                deleteFile(result, file.name);
+            })
+            .catch((ex: any) => {
+                console.log(ex);
+            });
     };
 
     const returnField = (fieldName) => {
         return <Typography variant="h6">{fieldName || '-'}</Typography>;
     };
 
+    const cancelAllDownloads = () => {
+        try {
+            controllerSas.abort();
+            controllerSas = new AbortController();
+        } catch (error) {
+            console.log(error);
+        }
+
+        abortArray.forEach((file: any) => {
+            try {
+                file.controller.abort();
+            } catch (ex) {
+                console.log(ex);
+            }
+        });
+    };
+
+    const checkIfDeleteIsEnabled = (_file): boolean => {
+        if (!dataset.permissions.editDataset) {
+            return true;
+        }
+        const index = progressArray.findIndex((x: any) => x.name === _file.name);
+        if (index === -1) {
+            return false;
+        }
+        if (progressArray[index].percent === 1) {
+            return true;
+        }
+        return false;
+    };
+
     return !showEditDataset ? (
-        !datasetResponse.loading && !dataset.id && datasetResponse.notFound ? (
+        !loadingFiles && !dataset.id && datasetResponse.notFound ? (
             <NotFound />
         ) : (
-            <OuterWrapper>
-                {loading && <LoadingFull noTimeout={datasetDeleteInProgress} />}
-                {userClickedDelete && (
-                    <DeleteResourceComponent
-                        ResourceName={dataset?.name}
-                        setUserClickedDelete={setUserClickedDelete}
-                        onClick={deleteDataset}
-                        type="dataset"
-                    />
-                )}
-                <Wrapper>
-                    <div>
-                        <div style={{ marginBottom: '16px' }}>
-                            <Typography variant="h1">{dataset?.name}</Typography>
-                            {!checkUrlIfGeneralDataset() ? (
-                                <span>This data is only available for this study</span>
-                            ) : null}
-                        </div>
-                        {!checkUrlIfGeneralDataset() ? (
-                            <Link
-                                to={'/studies/' + studyId}
-                                style={{ color: '#007079', fontSize: '22px', margin: '0 0 0 16px' }}
-                                data-cy="dataset_back_to_study"
-                            >
-                                <Icon color="#007079" name="arrow_back" size={24} style={{ marginRight: '16px' }} />
-                                Back to study
-                            </Link>
-                        ) : (
-                            <Link to={'/datasets'} style={{ color: '#007079', fontSize: '22px', margin: '0 0 0 16px' }}>
-                                <Icon color="#007079" name="arrow_back" size={24} style={{ marginRight: '16px' }} />
-                                Back to datasets
-                            </Link>
-                        )}
-                        <Dropzone
-                            onDrop={(event: File[]) => handleFileDrop(event)}
-                            disabled={
-                                !(
-                                    dataset.permissions?.editDataset &&
-                                    (percentComplete === 0 || percentComplete === 100)
-                                )
-                            }
+            <>
+                <Prompt
+                    hasChanged={hasChanged}
+                    fallBackAddress={!checkUrlIfGeneralDataset() ? '/studies/' + studyId : undefined}
+                    customText="All downloads will cancel"
+                />
+                <OuterWrapper>
+                    {loading && <LoadingFull noTimeout={datasetDeleteInProgress} />}
+                    {userClickedDelete && (
+                        <DeleteResourceComponent
+                            ResourceName={dataset?.name}
+                            setUserClickedDelete={setUserClickedDelete}
+                            onClick={deleteDataset}
+                            type="dataset"
                         />
-                        {percentComplete > 0 && (
-                            <div style={{ display: 'flex' }}>
-                                <LinearProgress
-                                    style={{ marginTop: '16px' }}
-                                    value={percentComplete}
-                                    variant="determinate"
-                                />
-                                <Button
-                                    onClick={() => {
-                                        source.cancel();
-                                        setPercentComplete(0);
-
-                                        cancelToken = axios.CancelToken;
-                                        source = cancelToken.source();
-                                    }}
-                                    style={{ float: 'right', padding: '4px' }}
-                                    variant="ghost_icon"
-                                    disabled={percentComplete === 0 || percentComplete === 100}
+                    )}
+                    <Wrapper>
+                        <div>
+                            <div style={{ marginBottom: '16px' }}>
+                                <Typography variant="h1">{dataset?.name}</Typography>
+                                {!checkUrlIfGeneralDataset() ? (
+                                    <span>This data is only available for this study</span>
+                                ) : null}
+                            </div>
+                            {!checkUrlIfGeneralDataset() ? (
+                                <Link
+                                    to={'/studies/' + studyId}
+                                    style={{ color: '#007079', fontSize: '22px', margin: '0 0 0 16px' }}
+                                    data-cy="dataset_back_to_study"
                                 >
-                                    {percentComplete === 0 || percentComplete === 100
-                                        ? EquinorIcon('check', '', 24)
-                                        : EquinorIcon('clear', '', 24)}
-                                </Button>
-                            </div>
-                        )}
-                        <div style={{ paddingTop: '8px' }}>
-                            {files &&
-                                files.map((file: any, i: number) => {
-                                    return (
-                                        <AttachmentWrapper key={getKey()}>
-                                            <div>{file.name}</div>
-                                            <div>{bytesToMB(file.size) + ' '} MB</div>
-                                            <Button
-                                                variant="ghost_icon"
-                                                onClick={() => removeFile(i, file)}
-                                                style={{ marginTop: '-8px' }}
-                                                disabled={!(percentComplete === 0 || percentComplete === 100)}
-                                            >
-                                                <Icon
-                                                    color="#007079"
-                                                    name="delete_forever"
-                                                    size={24}
-                                                    style={{ cursor: 'pointer' }}
-                                                />
-                                            </Button>
-                                        </AttachmentWrapper>
-                                    );
-                                })}
-                        </div>
-                    </div>
-                    {!datasetResponse.loading ? (
-                        <RightWrapper>
-                            <div>
-                                <Label>Storage account</Label>
-                                <a href={dataset?.storageAccountLink} target="_blank" rel="noopener noreferrer">
-                                    <span style={{ marginRight: '8px' }}>{dataset?.storageAccountName}</span>
-                                    {EquinorIcon('external_link', '#007079', 24)}
-                                </a>
-                            </div>
-                            <div>
-                                <Label>Location</Label>
-                                {returnField(dataset?.location)}
-                            </div>
-                            <div>
-                                <Label>Data classification</Label>
-                                {returnField(dataset?.classification)}
-                            </div>
-                            <div>
-                                <Label>LRA ID</Label>
-                                {returnField(dataset?.lraId)}
-                            </div>
-                            <div>
-                                <Label>Data ID</Label>
-                                {returnField(dataset?.dataId)}
-                            </div>
-                            <div>
-                                <Label>Source system</Label>
-                                {returnField(dataset?.sourceSystem)}
-                            </div>
-                            <div>
-                                <Label>BA data owner</Label>
-                                {returnField(dataset?.baDataOwner)}
-                            </div>
-                            <div>
-                                <Label>Asset</Label>
-                                {returnField(dataset?.asset)}
-                            </div>
-                            <div>
-                                <Label>Country of origin</Label>
-                                {returnField(dataset?.countryOfOrigin)}
-                            </div>
-                            <div>
-                                <Label>Area L1</Label>
-                                {returnField(dataset?.areaL1)}
-                            </div>
-                            <div>
-                                <Label>Area L2</Label>
-                                {returnField(dataset?.areaL2)}
-                            </div>
-                            <div>
-                                <div style={{ display: 'inline-block', marginRight: '8px' }}>
-                                    <Tooltip
-                                        title={
-                                            !(
-                                                permissions.canEdit_PreApproved_Datasets ||
-                                                dataset.permissions?.editDataset
-                                            )
-                                                ? 'You do not have permission to edit metadata'
-                                                : ''
-                                        }
-                                        placement="top"
+                                    <Icon color="#007079" name="arrow_back" size={24} style={{ marginRight: '16px' }} />
+                                    Back to study
+                                </Link>
+                            ) : (
+                                <Link
+                                    to="/datasets"
+                                    style={{ color: '#007079', fontSize: '22px', margin: '0 0 0 16px' }}
+                                >
+                                    <Icon color="#007079" name="arrow_back" size={24} style={{ marginRight: '16px' }} />
+                                    Back to datasets
+                                </Link>
+                            )}
+                            {!datasetResponse.loading && (
+                                <Dropzone
+                                    onDrop={(event: File[]) => handleFileDrop(event)}
+                                    loading={
+                                        dataset.storageAccountLink !== '' && dataset.storageAccountLink !== null
+                                            ? false
+                                            : true
+                                    }
+                                    disabled={!(dataset.permissions?.editDataset && dataset.storageAccountLink)}
+                                />
+                            )}
+                            {duplicateFiles && (
+                                <div>
+                                    <Chip
+                                        variant="active"
+                                        onDelete={() => {
+                                            setDuplicateFiles(false);
+                                        }}
+                                        style={{ marginLeft: 'auto' }}
                                     >
-                                        <Button
-                                            style={{ width: '150px' }}
-                                            variant="outlined"
-                                            onClick={handleEditMetdata}
-                                            data-cy="dataset_edit"
-                                            disabled={
-                                                !(
-                                                    permissions.canEdit_PreApproved_Datasets ||
-                                                    dataset.permissions?.editDataset
-                                                )
-                                            }
-                                        >
-                                            Edit metadata
-                                        </Button>
-                                    </Tooltip>
+                                        Already uploaded files are skipped
+                                    </Chip>
                                 </div>
-                                {!checkUrlIfGeneralDataset() && (
-                                    <div style={{ display: 'inline-block', marginTop: '8px' }}>
+                            )}
+                            <div style={{ paddingTop: '8px' }}>
+                                {!loadingFiles ? (
+                                    files.length > 0 ? (
+                                        files.map((file: any, i: number) => {
+                                            return (
+                                                <div key={file.name}>
+                                                    <AttachmentWrapper>
+                                                        <div>{file.name}</div>
+                                                        <div>{bytesToSize(file.size)} </div>
+                                                        <Button
+                                                            variant="ghost_icon"
+                                                            onClick={() => removeFile(i, file)}
+                                                            style={{ marginTop: '-8px' }}
+                                                            disabled={checkIfDeleteIsEnabled(file)}
+                                                        >
+                                                            <Icon
+                                                                color="#007079"
+                                                                name="delete_forever"
+                                                                size={24}
+                                                                style={{ cursor: 'pointer' }}
+                                                            />
+                                                        </Button>
+                                                    </AttachmentWrapper>
+                                                    {file.percent && (
+                                                        <LinearProgress
+                                                            style={{ marginBottom: '16px', marginTop: '-4px' }}
+                                                            value={file.percent}
+                                                            variant="determinate"
+                                                        />
+                                                    )}
+                                                </div>
+                                            );
+                                        })
+                                    ) : (
+                                        <div style={{ textAlign: 'center' }}>
+                                            {dataset.storageAccountLink ? 'No files uploaded yet.' : ''}
+                                        </div>
+                                    )
+                                ) : (
+                                    <div style={{ textAlign: 'center' }}>
+                                        <DotProgress color="primary" />
+                                        <div>Loading files..</div>
+                                    </div>
+                                )}
+                            </div>
+                        </div>
+                        {!datasetResponse.loading ? (
+                            <RightWrapper>
+                                <div>
+                                    <Label>Storage account</Label>
+                                    {dataset?.storageAccountLink ? (
+                                        <a href={dataset?.storageAccountLink} target="_blank" rel="noopener noreferrer">
+                                            <span style={{ marginRight: '8px' }}>{dataset?.storageAccountName}</span>
+                                            {EquinorIcon('external_link', '#007079', 24)}
+                                        </a>
+                                    ) : (
+                                        <Tooltip title={storageAccountStatus} placement="top">
+                                            <DotProgress color="primary" />
+                                        </Tooltip>
+                                    )}
+                                </div>
+                                <div>
+                                    <Label>Location</Label>
+                                    {returnField(dataset?.location)}
+                                </div>
+                                <div>
+                                    <Label>Data classification</Label>
+                                    {returnField(dataset?.classification)}
+                                </div>
+                                <div>
+                                    <Label>LRA ID</Label>
+                                    {returnField(dataset?.lraId)}
+                                </div>
+                                <div>
+                                    <Label>Data ID</Label>
+                                    {returnField(dataset?.dataId)}
+                                </div>
+                                <div>
+                                    <Label>Source system</Label>
+                                    {returnField(dataset?.sourceSystem)}
+                                </div>
+                                <div>
+                                    <Label>BA data owner</Label>
+                                    {returnField(dataset?.baDataOwner)}
+                                </div>
+                                <div>
+                                    <Label>Asset</Label>
+                                    {returnField(dataset?.asset)}
+                                </div>
+                                <div>
+                                    <Label>Country of origin</Label>
+                                    {returnField(dataset?.countryOfOrigin)}
+                                </div>
+                                <div>
+                                    <Label>Area L1</Label>
+                                    {returnField(dataset?.areaL1)}
+                                </div>
+                                <div>
+                                    <Label>Area L2</Label>
+                                    {returnField(dataset?.areaL2)}
+                                </div>
+                                <div>
+                                    <div style={{ display: 'inline-block', marginRight: '8px' }}>
                                         <Tooltip
                                             title={
                                                 !(
@@ -418,28 +686,59 @@ const DatasetDetails = (props: any) => {
                                             <Button
                                                 style={{ width: '150px' }}
                                                 variant="outlined"
-                                                color="danger"
-                                                onClick={() => setUserClickedDelete(true)}
-                                                data-cy="dataset_delete"
+                                                onClick={handleEditMetdata}
+                                                data-cy="dataset_edit"
                                                 disabled={
                                                     !(
                                                         permissions.canEdit_PreApproved_Datasets ||
-                                                        dataset.permissions?.deleteDataset
+                                                        dataset.permissions?.editDataset
                                                     )
                                                 }
                                             >
-                                                Delete data set
+                                                Edit metadata
                                             </Button>
                                         </Tooltip>
                                     </div>
-                                )}
-                            </div>
-                        </RightWrapper>
-                    ) : (
-                        <LoadingFull noTimeout={datasetDeleteInProgress} />
-                    )}
-                </Wrapper>
-            </OuterWrapper>
+                                    {!checkUrlIfGeneralDataset() && (
+                                        <div style={{ display: 'inline-block', marginTop: '8px' }}>
+                                            <Tooltip
+                                                title={
+                                                    !(
+                                                        permissions.canEdit_PreApproved_Datasets ||
+                                                        dataset.permissions?.editDataset
+                                                    )
+                                                        ? 'You do not have permission to edit metadata'
+                                                        : ''
+                                                }
+                                                placement="top"
+                                            >
+                                                <Button
+                                                    style={{ width: '150px' }}
+                                                    variant="outlined"
+                                                    color="danger"
+                                                    onClick={() => setUserClickedDelete(true)}
+                                                    data-cy="dataset_delete"
+                                                    disabled={
+                                                        !(
+                                                            (permissions.canEdit_PreApproved_Datasets &&
+                                                                checkUrlIfGeneralDataset()) ||
+                                                            dataset.permissions?.deleteDataset
+                                                        )
+                                                    }
+                                                >
+                                                    Delete data set
+                                                </Button>
+                                            </Tooltip>
+                                        </div>
+                                    )}
+                                </div>
+                            </RightWrapper>
+                        ) : (
+                            <LoadingFull noTimeout={datasetDeleteInProgress} />
+                        )}
+                    </Wrapper>
+                </OuterWrapper>
+            </>
         )
     ) : (
         <CreateEditDataset
@@ -447,7 +746,6 @@ const DatasetDetails = (props: any) => {
             setDatasetFromDetails={setDataset}
             setShowEditDataset={setShowEditDataset}
             editingDataset
-            cache={datasetResponse.cache}
             permissions={dataset.permissions}
         />
     );
