@@ -1,3 +1,4 @@
+/*eslint-disable consistent-return */
 import React, { useState, useEffect, useRef } from 'react';
 import { Table, Icon, Button, Checkbox, SideSheet, Tooltip } from '@equinor/eds-core-react';
 import { checkbox } from '@equinor/eds-icons';
@@ -8,6 +9,7 @@ import DatasetSearchFilter from './DatasetSearchFilter';
 import DatasetSidesheetView from './DatasetSidesheetView';
 import DropdownFilter from '../common/customComponents/DropdownFilter';
 import useClickOutside from '../common/customComponents/useClickOutside';
+import '../../styles/Table.scss';
 import Cookies from 'js-cookie';
 
 const { Body, Row, Cell, Head } = Table;
@@ -49,7 +51,7 @@ interface checkedColumns {
     sepesApproved: boolean;
 }
 
-interface filter {
+interface Filter {
     name: string;
     sourceSystem: string;
     areaL2: string;
@@ -95,7 +97,7 @@ const DatasetsOverviewTable: React.FC<DatasetsOverviewTableProps> = ({ datasets,
               }
     );
 
-    const [filter, setFilter] = useState<filter>({
+    const [filter, setFilter] = useState<Filter>({
         name: '',
         sourceSystem: '',
         areaL2: '',
@@ -133,11 +135,7 @@ const DatasetsOverviewTable: React.FC<DatasetsOverviewTableProps> = ({ datasets,
 
     const returnCell = (checker: any, fieldName?: string | number, header?: boolean) => {
         if (header) {
-            return checker ? (
-                <Cell as="th" scope="col">
-                    {fieldName || ''}
-                </Cell>
-            ) : null;
+            return checker ? <Cell scope="col">{fieldName || ''}</Cell> : null;
         }
         return checker ? <Cell>{fieldName || '-'}</Cell> : null;
     };
@@ -150,7 +148,6 @@ const DatasetsOverviewTable: React.FC<DatasetsOverviewTableProps> = ({ datasets,
                 name={name}
                 value={checked.toString()}
                 onChange={handleColumnsChange}
-                enterKeyHint="Select filter"
             />
         );
     };
@@ -192,7 +189,7 @@ const DatasetsOverviewTable: React.FC<DatasetsOverviewTableProps> = ({ datasets,
 
     const filterListOptions = (column: string, filterColumn: any, resDataset) => {
         if (filterColumn.length > 0) {
-            let res = resDataset;
+            const res = resDataset;
             return combineArray(
                 filterColumn.map((row: any) => {
                     return filterList(column, row, res);
@@ -203,7 +200,7 @@ const DatasetsOverviewTable: React.FC<DatasetsOverviewTableProps> = ({ datasets,
     };
 
     function combineArray(array: any) {
-        let newArray: any = [];
+        const newArray: any = [];
         array.map((res: any, i: number) => {
             res.map((dataset: any, j: number) => {
                 newArray.push(dataset);
@@ -320,7 +317,7 @@ const DatasetsOverviewTable: React.FC<DatasetsOverviewTableProps> = ({ datasets,
                     </Row>
                     {datasets &&
                         applyFilter().map((row: DatasetObj) => (
-                            <Row key={row.id} onClick={() => handleOnclick(row)} style={{ cursor: 'pointer' }}>
+                            <Row key={row.id} onClick={() => handleOnclick(row)} id="tableRow">
                                 {returnCell(checkedColums.name, row.name)}
                                 {returnCell(checkedColums.sourceSystem, row.sourceSystem)}
                                 {returnCell(checkedColums.areaL2, row.areaL2)}
