@@ -1,21 +1,21 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Typography, Icon } from '@equinor/eds-core-react';
-import { Link } from 'react-router-dom';
-import { lock, lock_open } from '@equinor/eds-icons';
+import { visibility, visibility_off } from '@equinor/eds-icons';
 import CustomLogoComponent from '../common/CustomLogoComponent';
 import { StudyObj } from '../common/interfaces';
+import { useHistory } from 'react-router-dom';
 
 const icons = {
-    lock,
-    lock_open
+    visibility,
+    visibility_off
 };
 Icon.add(icons);
 
 const SmallText = styled.div`
     font-size: 10px;
     display: inline-block;
-    width: 72px;
+    width: 76px;
 `;
 
 const Wrapper = styled.div`
@@ -26,6 +26,7 @@ const Wrapper = styled.div`
     padding: 16px;
     min-width:120px;
     margin: 0 0px 16px 32px;
+    cursor: pointer;
     background-color: #ffffff;
     @media (max-width: 385px) {
       grid-template-columns: minmax(116px,200px) 1fr;
@@ -64,26 +65,30 @@ type StudyComponentProps = {
 const StudyComponent: React.FC<StudyComponentProps> = ({ study }) => {
     const { name, description, restricted, id, vendor, logoUrl } = study;
     const url = '/studies/' + id;
+    const history = useHistory();
 
     return (
-        <Link to={url} style={{ color: 'black', textDecoration: 'none' }}>
-            <Wrapper>
-                <LogoTitleWrapper>
-                    <CustomLogoComponent logoUrl={logoUrl} />
+        <Wrapper onClick={() => history.push({ pathname: url, state: { userCameFromHome: true } })}>
+            <LogoTitleWrapper>
+                <CustomLogoComponent logoUrl={logoUrl} center />
+                <div>
+                    <Typography variant="h6">{name}</Typography>
+                    <SmallText>{vendor}</SmallText>
                     <div>
-                        <Typography variant="h6">{name}</Typography>
-                        <SmallText>{vendor}</SmallText>
-                        <div>
-                            <SmallText>
-                                {restricted ? 'Hidden' : 'Not hidden'}
-                                <Icon color="#007079" name={restricted ? 'lock' : 'lock_open'} size={24} />
-                            </SmallText>
-                        </div>
+                        <SmallText>
+                            {restricted ? 'Hidden' : 'Not hidden'}
+                            <Icon
+                                color="#007079"
+                                name={restricted ? 'visibility_off' : 'visibility'}
+                                size={16}
+                                style={{ marginLeft: '4px' }}
+                            />
+                        </SmallText>
                     </div>
-                </LogoTitleWrapper>
-                <div>{description}</div>
-            </Wrapper>
-        </Link>
+                </div>
+            </LogoTitleWrapper>
+            <div>{description}</div>
+        </Wrapper>
     );
 };
 
