@@ -41,6 +41,8 @@ type SandboxConfigProps = {
     permissions: SandboxPermissions;
     setSandbox: any;
     sandbox: SandboxObj;
+    getResources: any;
+    controller: AbortController;
 };
 
 const SandboxConfig: React.FC<SandboxConfigProps> = ({
@@ -50,7 +52,9 @@ const SandboxConfig: React.FC<SandboxConfigProps> = ({
     setUpdateCache,
     permissions,
     setSandbox,
-    sandbox
+    sandbox,
+    getResources,
+    controller
 }) => {
     return (
         <Wrapper>
@@ -62,11 +66,12 @@ const SandboxConfig: React.FC<SandboxConfigProps> = ({
                     permissions={permissions}
                     setSandbox={setSandbox}
                     sandbox={sandbox}
+                    controller={controller}
                 />
                 <PolictyComponentWrapper>
-                    <PolicyComponent displayCheckbox />
+                    <PolicyComponent displayCheckbox sandbox={sandbox} />
                 </PolictyComponentWrapper>
-                <ResourcesComponent resources={resources} />
+                <ResourcesComponent resources={resources} getResources={getResources} />
             </InfoWrapper>
         </Wrapper>
     );
