@@ -1,38 +1,40 @@
+/* eslint-disable react/no-array-index-key */
 import React from 'react';
 import { Table } from '@equinor/eds-core-react';
 import ResourceItemComponent from './ResourceItemComponent';
+import '../../../styles/Table.scss';
 
 const { Body, Row, Cell, Head } = Table;
 
 const Dataset = (props: any) => {
-    const { resources } = props;
+    const { resources, getResources } = props;
     return (
         <Table style={{ width: '100%', marginBottom: '24px' }}>
             <Head>
                 <Row>
-                    <Cell as="th" scope="col">
-                        Resources
-                    </Cell>
+                    <Cell scope="col">Resources</Cell>
                 </Row>
             </Head>
             <Body>
-                {resources ? (
+                {resources && Array.isArray(resources) ? (
                     resources.map((resource: any, i: number) => {
                         return (
-                            <Row key={i}>
+                            <Row key={i} id="tableRowNoPointerNoColor">
                                 <Cell>
                                     <ResourceItemComponent
                                         name={resource.name}
                                         type={resource.type}
                                         status={resource.status}
                                         linkToResource={resource.linkToExternalSystem}
+                                        retryLink={resource.retryLink}
+                                        getResources={getResources}
                                     />
                                 </Cell>
                             </Row>
                         );
                     })
                 ) : (
-                    <Row key={1}>
+                    <Row key={1} id="tableRowNoPointerNoColor">
                         <Cell>No resources...</Cell>
                     </Row>
                 )}
