@@ -24,9 +24,9 @@ const MoreActionsWrapper = styled.div`
     position: absolute;
     background-color: #ffffff;
     box-shadow: 0 0 4px 4px #e7e7e7;
-    width: 240px;
+    width: 296px;
     border-radius: 4px;
-    margin-top: 40px;
+    margin-top: 196px;
     display: grid;
     grid-template-rows: 1fr 1fr;
 `;
@@ -86,12 +86,12 @@ const VmProperties: React.FC<VmPropertiesProps> = ({
     const deleteVm = (): void => {
         setUpdateCache({ ...updateCache, [getVmsForSandboxUrl(sandboxId)]: true });
         setActiveTab(0);
-        let currentVms: any = [...vms];
+        const currentVms: any = [...vms];
         currentVms.splice(vms.indexOf(vmProperties), 1);
         setVms(currentVms);
         deleteVirtualMachine(vmProperties.id).then((result: any) => {
-            if (result.Message) {
-                notify.show('danger', '500', result.Message, result.RequestId);
+            if (result && result.Message) {
+                notify.show('danger', '500', result);
             } else {
                 getResources();
             }
@@ -132,9 +132,9 @@ const VmProperties: React.FC<VmPropertiesProps> = ({
                             }
                             placement="top"
                         >
-                            <Button style={{ width: '216px' }} disabled={!vmProperties.linkToExternalSystem}>
-                                Open virtual machine
-                                <div style={{ marginLeft: 'auto', display: 'block' }}>
+                            <Button style={{ width: '296px' }} disabled={!vmProperties.linkToExternalSystem}>
+                                Connect to virtual machine
+                                <div style={{ marginLeft: 'auto', paddingLeft: '39px' }}>
                                     {EquinorIcon('external_link', '#FFFFFF', 24, () => {}, true)}
                                 </div>
                             </Button>
@@ -144,12 +144,12 @@ const VmProperties: React.FC<VmPropertiesProps> = ({
 
                 <Button
                     variant="outlined"
-                    style={{ marginTop: '8px', width: '216px' }}
+                    style={{ marginTop: '8px', width: '296px' }}
                     onClick={() => handleToggle()}
                     data-cy="vm_more_actions"
                 >
                     More actions
-                    <div style={{ marginLeft: 'auto' }}>
+                    <div style={{ marginLeft: 'auto', paddingLeft: '39px' }}>
                         {EquinorIcon('arrow_drop_down', '#007079', 24, () => {}, true)}
                     </div>
                     {displayMoreActions && (
@@ -164,20 +164,27 @@ const VmProperties: React.FC<VmPropertiesProps> = ({
                                 {EquinorIcon('key', '#6F6F6F', 24, () => {}, true)}
                                 <ItemText>Reset password</ItemText>
                             </Item>
-                            <Item
-                                color="#EB0000"
-                                style={{
-                                    opacity: permissions.update ? 1 : 0.5,
-                                    pointerEvents: permissions.update ? 'initial' : 'none'
-                                }}
-                                onClick={() => {
-                                    setUserClickedDelete(true);
-                                }}
-                                data-cy="vm_delete"
+                            <Tooltip
+                                title={permissions.update ? '' : 'You do not have access to delete VMs'}
+                                placement="right"
+                                open={!permissions.update}
                             >
-                                {EquinorIcon('delete_forever', '#EB0000', 24, () => {}, true)}
-                                <ItemText>Delete virtual machine</ItemText>
-                            </Item>
+                                <Item
+                                    color="#EB0000"
+                                    style={{
+                                        opacity: permissions.update ? 1 : 0.5,
+                                        pointerEvents: permissions.update ? 'initial' : 'none',
+                                        width: '296px'
+                                    }}
+                                    onClick={() => {
+                                        setUserClickedDelete(true);
+                                    }}
+                                    data-cy="vm_delete"
+                                >
+                                    {EquinorIcon('delete_forever', '#EB0000', 24, () => {}, true)}
+                                    <ItemText>Delete virtual machine</ItemText>
+                                </Item>
+                            </Tooltip>
                         </MoreActionsWrapper>
                     )}
                 </Button>
