@@ -1,17 +1,16 @@
-import React, { useState, useContext } from 'react';
-import { Button, Tooltip } from '@equinor/eds-core-react';
+import React from 'react';
+import { Button, Tooltip, Icon } from '@equinor/eds-core-react';
 import styled from 'styled-components';
 import { close } from '@equinor/eds-icons';
-import { Icon } from '@equinor/eds-core-react';
-import { Link, useHistory } from 'react-router-dom';
-import { getDatasetList, addStudyDataset, removeStudyDataset, unlinkStudyDataset } from '../../services/Api';
+import { useHistory } from 'react-router-dom';
+import { unlinkStudyDataset } from '../../services/Api';
 import { StudyObj } from '../common/interfaces';
-import SearchWithDropdown from '../common/customComponents/SearchWithDropdown';
+//import SearchWithDropdown from '../common/customComponents/SearchWithDropdown';
 import DatasetsTable from './Tables/DatasetsTable';
 import * as notify from '../common/notify';
-import { Permissions } from '../../index';
-import useFetchUrl from '../common/hooks/useFetchUrl';
-import { getDatasetsInStudyUrl, getDatasetsUrl, getStudyByIdUrl } from '../../services/ApiCallStrings';
+//import { Permissions } from '../../index';
+//import useFetchUrl from '../common/hooks/useFetchUrl';
+import { getDatasetsInStudyUrl, getStudyByIdUrl } from '../../services/ApiCallStrings';
 
 const icons = {
     close
@@ -20,7 +19,7 @@ Icon.add(icons);
 
 const Wrapper = styled.div`
     display: grid;
-    grid-template-rows: 20px minmax(299px, 1fr);
+    grid-template-rows: 20px minmax(301px, 1fr);
     grid-gap: 23px;
 `;
 // Might have to change back when general datasets is added back
@@ -32,7 +31,7 @@ const TableWrapper = styled.div`
         margin-top: 64px;
     }
 `;
-
+/*
 const Bar = styled.div`
     display: grid;
     grid-template-columns: 1fr 0.3fr 296px;
@@ -48,6 +47,7 @@ const Bar = styled.div`
         grid-gap: 8px;
     }
 `;
+*/
 
 type StudyComponentFullProps = {
     study: StudyObj;
@@ -58,10 +58,10 @@ type StudyComponentFullProps = {
 
 const DataSetComponent: React.FC<StudyComponentFullProps> = ({ study, setStudy, setUpdateCache, updateCache }) => {
     const history = useHistory();
-    const [datasetsList, setDatasetsList] = useState<any>([]);
-    const [isOpen, setIsOpen] = useState<boolean>(false);
-    const permissions = useContext(Permissions);
-    const datasetsResponse = useFetchUrl(getDatasetsUrl(), setDatasetsList, permissions.canRead_PreApproved_Datasets);
+    //const [datasetsList, setDatasetsList] = useState<any>([]);
+    //const [isOpen, setIsOpen] = useState<boolean>(false);
+    //const permissions = useContext(Permissions);
+    //const datasetsResponse = useFetchUrl(getDatasetsUrl(), setDatasetsList, permissions.canRead_PreApproved_Datasets);
     const removeDataset = (row: any) => {
         const studyId = window.location.pathname.split('/')[2];
         setStudy({ ...study, datasets: study.datasets.filter((dataset: any) => dataset.id !== row.id) });
@@ -72,10 +72,10 @@ const DataSetComponent: React.FC<StudyComponentFullProps> = ({ study, setStudy, 
         });
         unlinkStudyDataset(studyId, row.id).then((result: any) => {
             if (result && result.Message) {
-                notify.show('danger', '500', result.Message, result.RequestId);
+                notify.show('danger', '500', result);
                 console.log('Err');
             }
-            datasetsResponse.setLoading(false);
+            //datasetsResponse.setLoading(false);
         });
     };
 
@@ -89,7 +89,7 @@ const DataSetComponent: React.FC<StudyComponentFullProps> = ({ study, setStudy, 
             }
         });
     };
-
+    /*
     const addDatasetToStudy = (row: any) => {
         setUpdateCache({ ...updateCache, [getStudyByIdUrl(study.id)]: true, [getDatasetsInStudyUrl(study.id)]: true });
         setIsOpen(false);
@@ -100,14 +100,13 @@ const DataSetComponent: React.FC<StudyComponentFullProps> = ({ study, setStudy, 
             setStudy({ ...study, datasets: datasetList });
             addStudyDataset(studyId, row.id).then((result: any) => {
                 if (result && result.Message) {
-                    notify.show('danger', '500', result.Message, result.RequestId);
+                    notify.show('danger', '500', result);
                     console.log('Err');
                 }
                 datasetsResponse.setLoading(false);
             });
         }
     };
-
     const checkIfDatasetIsAlreadyAdded = (id: string) => {
         let elementExist = false;
         study.datasets &&
@@ -118,7 +117,7 @@ const DataSetComponent: React.FC<StudyComponentFullProps> = ({ study, setStudy, 
             });
         return elementExist;
     };
-
+*/
     return (
         <Wrapper>
             {/*<Bar>*/}
@@ -143,7 +142,7 @@ const DataSetComponent: React.FC<StudyComponentFullProps> = ({ study, setStudy, 
                     </Button>
                 </Tooltip>
             </div>
-            {/* 
+            {/*
                 <span style={{ textAlign: 'center' }}>or</span>
                 <div onMouseEnter={() => setIsOpen(true)} onMouseLeave={() => setIsOpen(false)}>
                     <SearchWithDropdown
@@ -155,7 +154,6 @@ const DataSetComponent: React.FC<StudyComponentFullProps> = ({ study, setStudy, 
                         disabled={study.permissions && !study.permissions.addRemoveDataset}
                     />
                 </div>
-                
             </Bar>
             <Link to="/datasets" style={{ color: '#007079', float: 'right', marginLeft: 'auto', marginTop: '32px' }}>
                 Advanced search
