@@ -1,7 +1,9 @@
+/* eslint-disable react/jsx-curly-brace-presence */
 import React from 'react';
 import { Table, Icon } from '@equinor/eds-core-react';
 import { chevron_right } from '@equinor/eds-icons';
-import { Link } from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
+import '../../../styles/Table.scss';
 
 const { Body, Row, Cell, Head } = Table;
 const icons = {
@@ -15,28 +17,24 @@ type SandboxTableProps = {
 
 const SandboxTable: React.FC<SandboxTableProps> = ({ sandboxes }) => {
     const studyId = window.location.pathname.split('/')[2];
+    const history = useHistory();
     const returnCell = (value: string, sandboxId: string, type: 'icon' | 'text') => {
         //This means it is a study specific dataset
-        return (
-            <Cell>
-                <Link
-                    style={{ color: '#000000', cursor: 'pointer' }}
-                    to={'/studies/' + studyId + '/sandboxes/' + sandboxId}
-                >
-                    {type === 'icon' ? <Icon name="chevron_right" size={24} /> : value}
-                </Link>
-            </Cell>
-        );
+        return <Cell>{type === 'icon' ? <Icon name="chevron_right" size={24} /> : value}</Cell>;
+    };
+
+    const redirectOnCellClick = (row: any) => {
+        if (studyId && row.id) {
+            history.push('/studies/' + studyId + '/sandboxes/' + row.id);
+        }
     };
     return (
         <div>
             <Table style={{ width: '100%', marginBottom: '24px' }}>
                 <Head>
                     <Row>
-                        <Cell as="th" scope="col">
-                            Sandbox
-                        </Cell>
-                        <Cell style={{ width: '10px' }} as="th" scope="col">
+                        <Cell scope="col">Sandbox</Cell>
+                        <Cell style={{ width: '10px' }} scope="col">
                             {''}
                         </Cell>
                     </Row>
@@ -44,13 +42,13 @@ const SandboxTable: React.FC<SandboxTableProps> = ({ sandboxes }) => {
                 <Body>
                     {sandboxes && sandboxes.length > 0 ? (
                         sandboxes.map((row) => (
-                            <Row key={row.id}>
+                            <Row key={row.id} id="tableRow" onClick={() => redirectOnCellClick(row)}>
                                 {returnCell(row.name, row.id, 'text')}
                                 {returnCell('', row.id, 'icon')}
                             </Row>
                         ))
                     ) : (
-                        <Row key={1}>
+                        <Row key={1} id="tableRowNoPointerNoColor">
                             <Cell>No sandboxes added</Cell>
                             <Cell>{''}</Cell>
                         </Row>
